Barotropic weight vectorization
This patch modifies the calculation of barotropic weight functions
wt_[uv] to a more vectorizable form.  The if-blocks used to evaluate the
piecewise function was replaced with a sequence of min/max evaluations,
which are vectorizable.

A nondimensional subroundoff term is also introducted to address the
division by zero in the evaluation of 1 - 0.5 * Instep / visc_rem.
When visc_rem is zero, this term will never be selected by the max()
operator, since it is negatively asymptotic.  When visc_rem is
sufficiently large to accept this term, the value of subroundoff is too
small to modify the significand of the floating point value.

In our runs, the overall time of btstep was reduced by 6%.

Experiment: benchmark
Platform: AMD Ryzen 5 2600
Compiler: gfortran 12.2
Flags: -O3 -mavx

Timer: (Ocean BT pre-calcs only)

Before:
- 1.781952
- 1.787916
- 1.790447

After:
- 1.544491
- 1.553111
- 1.544491

Timer: (Ocean barotropic mode stepping)

Before:
- 4.042281
- 4.050225
- 4.051413

After:
- 3.785838
- 3.822393
- 3.835511

Speedup:
~13% of "(Ocean BT pre-calc step)"
~6% speedup of btstep