Generic I/O layer for infra and netCDF
This patch introduces a generalized I/O interface, through a new class,
`MOM_file`, for reading and writing of axis-based model fields.

A common API for interacting with files is defined in the `MOM_file`
class, and two implementations are provided:

* `MOM_infra_file`, using the "infra" framework (i.e. FMS)
* `MOM_netcdf_file`, using native netCDF framework

This separation allows us to define certain generic functions by class,
and platform-specific functions by type.  It will also allow for removal
of legacy FMS1 operations which are no longer used for the majority of
I/O but are still required for isolated MOM files.

This change will allow us to move incompatible files from the FMS to the
netCDF implementation, while still preserving a common structure for
both files.

The majority of the details of `MOM_file` and its subclasses are defined
in `MOM_io_file.F90`, which is exclusively accessed through
`MOM_io.F90`.  The netCDF implementation, designed to be used by
`MOM_netcdf_file` but is designed as a standalone system, is defined in
`MOM_netcdf.F90`.

*Interface*

`MOM_file` includes the following functions:

* `open`
* `close`
* `flush`
* `register_axis`
* `register_field`
* `write_attribute`
* `write_field`
* `file_is_open`
* `get_file_info`
* `get_file_fields`
* `get_field_atts`
* `read_field_chksum`

Most are designed to resemble the existing MOM I/O operations.

Note that some of these have not yet been implemented for
`MOM_netcdf_file`, since they are never used in the model.

`MOM_file_infra` includes the following additional operations:

* `get_file_times`
* `get_file_fieldtypes`

See documentation for usage of these functions.

The "axis"/"field" model from FMS1 has been preserved, where axes are
associated with variables which contain the grid point, and fields are
defined with respect to the file's axes.  Operations such as field
counters will exclude any variables associated with the axes.

The `axistype` and `fieldtype` from FMS have been replaced with new
abstractions (`MOM_axis` and `MOM_field`).  Internally, these point to
either the FMS types or equivalent netCDF types.

*Implementation*

Each file type contains an instance of its native type, as well as lists
of associated axes and fields.  List are implemented as linked lists of
names (stored as `label`) which are used to identify and extract or
write the axis/field to the internal type.  The mechanics of this are
largely hidden from the user and can be changed in the future, if
needed, without disruption to the rest of the codebase.

The current netCDF implementation very closely mirrors the FMS infra
behavior, but this can be relaxed or modified as needed.

netCDF I/O is currently only designed for serial I/O, with all of the
data on the root PE.  Further development would be needed to support
any kind of parallel I/O.

*Current Usage*

Two functions have been transferred from infra to the netCDF I/O:

* `Depth_list.nc`
* `ocean.stats.nc`

The following legacy functions and types from FMS have been preserved:

* `create_file`
* `reopen_file`
* `file_type`
* `open_file`
* `get_file_info`
* `get_file_fields`
* `get_file_times`

This is primarily to preserve compatibility with SIS2, but may also be
useful for other code, such as model drivers.  These may be phased out
in the future, however.