+Rossby_front_2d_init and BFB_init cleanup
  Completed the dimensional rescaling of internal variables in the
Rossby_front_initialization and BFB_initialization modules, including the
addition of comments describing numerous internal variables and their units. The
BFB module was more extensively modified to place logging of the module name and
version before the get_param calls for this module, following the MOM6 pattern,
to correct spelling errors in get_param descriptions, and to use the grid mask
in the grid file rather than a reexamination of the minimum depth to determine
the land-mask.  The internal subroutine write_BFB_log is no longer needed and
has been folded into BFB_set_coord.  All answers are bitwise identical, but
there are minor changes in the MOM_parameter_doc files for the buoy_forced_basin
test case.