Document units in 8 lat_param or diag modules
  Documented numerous internal variables and their units in 3 lateral
parameterization modules (MOM_hor_visc, MOM_lateral_mixing_coeffs,
MOM_MEKE_types) and the write_cputime module.  In addition, the units of a
number of internal variables in the MOM_sum_output module and arguments to
call_tracer_stocks were added to their descriptions in comments.  This commit
includes the addition of units arguments in 9 unlogged get_param calls, comments
describing why parameters are not rescaled, and the rescaling of MAXVEL in
hor_visc_init.  A number of spelling errors were also corrected in comments.
All answers and output are bitwise identical.