Scale default parameters
  Apply dimensional rescaling of parameters that are only used as defaults for
other parameters, with the rescaling explicitly undone as a part of the default
argument specification.  This includes de-emphasizing the use of GV%Angstrom_m,
instead using GV%Angstrom_Z.  As a result, fewer get_param calls read in
dimensional variables without specifying a scaling factor that can be compared
with their declared units.  All answers and output are bitwise identical.