(*)Correct two rescaling bugs in MOM_EOS
  Corrected the pressure rescaling in calculate_density_second_derivs_scalar()
that had been using an uninitialized variable.  Also corrected the dimensional
rescaling of dSVdT and dSVdS in calc_spec_vol_derivs_1d when temperature and
salinity are being rescaled but density is not.  Fortunately these do not seem
to impact solutions in any production runs, and there do not appear to be any
calls to calculate_density_second_derivs_scalar().  Also added checksum calls
for tv%varT, tv%varS and tv%covarTS to MOM_thermo_chksum when these elements of
the thermovar type are associated.  Comments were also added describing the
units of a number of internal variables or conversion factor in the MOM_EOS
module.  All answers in the MOM6-examples test suite are bitwise identical.