(*)Correct scaling of USE_STANLEY_PGF diagnostics
  Corrected the diagnostics rho_pgf, rho_stanley_pgf and p_stanley to only be
calculated if they would be written, give identical output when dimensional
rescaling is applied, and be documented with the right units.  Also added a
test for cases when USE_STANLEY_GM is set to true but STANLEY_COEF is negative
to reset the internal versions of this flag to false with a sensible warning
message rather than encountering segmentation faults.  All solutions are bitwise
identical in cases that worked before, but there are changes in some diagnostics
when they are dimensionally rescaled.