+Correct MOM_wave_interface unit descriptions
  Corrected the descriptions of 12 subroutine argument and internal variables in
MOM_wave_interface, including those of 4 arguments to Stokes_PGF, and those of 8
variables related to the Waves%ddt_Us_[xy] diagnostics, which are only enabled
with the STOKES_DDT flag that is labeled in the code as "developmental".  This
commit also includes the addition of the correct conversion arguments to the
register_diag_field calls for dudt_Stokes and dvdt_Stokes diagnostics, and
to the register_restart calls for Us_x_prev and Us_y_prev.  This also required
the addition of a unit_scale_type argument to waves_register_restarts.  All
solutions are bitwise identical, and the dimensional rescaling of two
diagnostics are corrected.