Rescale diagnostics from MOM_CVMix_KPP
  Rescaled 3-d diagnostics in MOM_CVMix_KPP, and do even more of the preliminary
calculations in rescaled variables before explicitly recasting variables into
MKS units for the calls to the CVMix routines.  With this change, all
register_diag_field routines in this module (apart from nondimensional
diagnostics) have conversion factors that can be tested for correctness can be
compared with the declared units of diagnostics.  All answers are bitwise
identical.