Made STANLEY_COEFF<0 warnings be fatal
We were issuing a WARNING when the user indicates to use the STANLEY
contribution to density but if the coefficient is negative we were
toggling the logical flag but only issuing a warning. This would lead
to users getting a result there were not expecting. Better for the
run to fail so they can set the coefficient to non-negative.