POSIX: siglongjmp and sigsetjmp_missing fixes
This patch fixes two issues in the POSIX API.

The `siglongjmp` interface was referencing the wrong symbol (`longjmp`).
While this did not seem to cause any issues, possibly due to some shared
definitions on glibc/BSD platforms, the error was correctly detected by
the Cray compiler.   This patch corrects the C symbol name.

The `sigsetjmp_missing` function, as a default replacement for a missing
`sigsetjmp`, was also defined without a return value, since it always
returns an error if called at runtime.  The Cray compiler raised a
warning about this, so we now assign a return value of -1, although it
is never used.

Thanks to Jim Edwards for reporting these errors.