+Add runtime parameter MEKE_LSCALE_MAX_VAL
  Added the runtime parameters MEKE_LSCALE_MAX_VAL and MEKE_MIN_DEPTH_TOT to
specify two previously hard-coded dimensional parameters that are used with some
options of the MEKE parameterizations.  Also added or amended comments
annotating the units of some of the internal variables in the MEKE module.  By
default, all answers are bitwise identical, but there are new entries in the
MOM_parameter_doc.all files for some configurations using the MEKE
parameterization.