+Add runtime parameter ENTRAIN_DIFFUSIVE_MAX_ENT
  Added the new runtime parameter ENTRAIN_DIFFUSIVE_MAX_ENT to specify a
previously hard-coded upper limit on the rate of entrainment in the buffer
layers with a bulk mixed layer in the MOM_entrainment_diffusive module.
By default all answers are bitwise identical, but there is a new entry in the
MOM_parameter_doc.all files for some layer-mode configurations with a bulk
mixed layer.