+Make Update_Stokes_Drift more robust
  Modified Update_Stokes_Drift to use more robust expressions when
WAVE_INTERFACE_ANSWER_DATE >= 20230101.  This new option may not be as fully
tested as it should be, but it appears to give answers that are similar to but
more robust than the previous expressions.  By default all answers are bitwise
identical, but there is a new use of the runtime parameter
WAVE_INTERFACE_ANSWER_DATE.