+Add runtime parameter REG_SFC_SUFFICIENT_ADJ
  Added the new runtime parameter REG_SFC_SUFFICIENT_ADJ to specify a previously
hard-coded fraction of the target net entrainment to the mixed and buffer layers
that is enough to stop the search for additional mass from the interior layers
when regularizing the near-surface layers in layer-mode configurations.  By
default all answers are bitwise identical, but there is a new entry in the
MOM_parameter_doc.all files for some layer-mode configurations with
REGULARIZE_SURFACE_LAYERS=True.