+Add runtime parameter SALT_EXTRACTION_LIMIT
  Added the new runtime parameter SALT_EXTRACTION_LIMIT to specify a previously
hard-coded limit on the fraction of the salt in a layer that can be extracted by
the surface fluxes within a timestep.  Also added do_not_log flags based on the
value of ENABLE_THERMODYNAMICS to avoid logging parameters that are only used if
thermodynamics are active in cases when it is not.  In addition, the doxygen
trailer describing this module had been inappropriately borrowed from another
module, so it was completely rewritten to describe what the routines in this
module actually do.  The missing doxygen description of set_pen_shortwave was
also added, and a number of spelling errors in comments were corrected.  By
default all answers are bitwise identical, but there is a new entry in the
MOM_parameter_doc.all files for some configurations and meaningless entries have
been removed from others.